ButtonBoardManager integration
We *still* weren't getting the correct April Tag ID passed to the DockWithAprilTagCommand due to the way the button bindings are generated at robot initialization. Had to create an overloaded constructor in the DockWithAprilTagCommand class to pass the ButtonBoardManager instance into it. Then, from the DockWithAprilTagCommand class, call the getAprilTagID method of the ButtonBoardManager class to get the correct ID. Crazy stupid workaround to get these button bindings to work correctly.