ButtonBox V2 Support and Pneumatics
Made changes to support the V2 of the ButtonBox. Checks out on Pi. Might have to adjust angular velocity constant for Rotate Right and Rotate Left buttons. I was only able to check these out on the cart and running on the floor might require a different value.  Also reversed the logic in the PneumaticSubsystem for open/close gripper settings. The new setup that was just installed is reversed from the previous one.